cloud_run/cv_analyzer/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
openai~=1.13.3
pymupdf
pydantic~=2.6.3
google-cloud-storage
google-cloud-firestore
Keras==2.15.0
numpy
scikit-learn
tensorflow==2.15.0
